separando publicaciones de feed ,agregando estilos
