+++ src/components/Profile.js
@@ -3,84 +3,24 @@ import { getAuth, onAuthStateChanged } from 'https://www.gstatic.com/firebasejs/
 import {
   collection, query, where, getDocs, getFirestore,
 } from 'https://www.gstatic.com/firebasejs/9.6.9/firebase-firestore.js';
-import { onNavigate } from '../main.js';
 import { headerTemplate } from './Header.js';
-import { dataPublication, reviewResultPublication } from '../cloudFirebase.js';
+import { publicationBeforeTemplate } from './PublicationBefore.js';
+import { publications } from './Publication.js';
 
 export const db = getFirestore();
 export const Profile = () => {
   const profileContainer = document.createElement('div');
   profileContainer.className = 'container-profile'; // contenedor general
-  const profileTemplate = document.createElement('main');
-  profileTemplate.className = 'container-publication';
-  const textProfile = document.createElement('p');
-  textProfile.innerText = 'Bienvenido a tu perfil';
-  textProfile.addEventListener('click', () => onNavigate('/'));
+  // agregando las publicaciones
+  const mainTemplate = document.createElement('main');
+  mainTemplate.className = 'container-publication';
+  // agregando las publicaciones al main
+  mainTemplate.appendChild(publicationBeforeTemplate());
+  mainTemplate.appendChild(publications());
 
-  /* PUBLICACIÓN DE USUARIO */
-  const sectionPublication = document.createElement('section');
-  sectionPublication.className = 'container-publication-section';
-  const figureSection = document.createElement('figure');
-  // foto de usuario
-  const imgPhotoUser = document.createElement('img');
-  imgPhotoUser.className = 'photo-user';
-  imgPhotoUser.src = 'img/profile-user.png';
-  imgPhotoUser.alt = 'foto de perfil';
-  const figcaptionUser = document.createElement('figcaption');
-  figcaptionUser.innerText = 'Username';
-  // inputs de publicación
-  const inputTitle = document.createElement('input');
-  inputTitle.placeholder = 'Titulo de publicación';
-  const inputText = document.createElement('textarea');
-  inputText.placeholder = 'Escriba su texto aqui';
-  inputText.className = 'input-text-publication';
-  // logos de publicación
-  const containerLogosButton = document.createElement('div');
-  const imgShareImage = document.createElement('img');
-  imgShareImage.className = 'share-image-logo logo-publication';
-  imgShareImage.src = 'img/insertar-icono-de-imagen.png';
-  imgShareImage.alt = 'logo para agregar imagenes a la publicación';
-
-  const imgShareStickers = document.createElement('img');
-  imgShareStickers.className = 'share-stickers-logo logo-publication';
-  imgShareStickers.src = 'img/emoticon-sonrisa.png';
-  imgShareStickers.alt = 'logo para agregar stickers a la publicación';
-
-  const imgTrash = document.createElement('img');
-  imgTrash.className = 'share-trash-logo logo-publication';
-  imgTrash.src = 'img/icons8-trash-30.png';
-  imgTrash.alt = 'logo para eliminar publicación';
-
-  const buttonPublication = document.createElement('button');
-  buttonPublication.className = 'button-publication';
-  buttonPublication.innerText = 'Publicar';
-
-  // agregando contenedores pequeños a medianos
-  figureSection.appendChild(imgPhotoUser);
-  figureSection.appendChild(figcaptionUser);
-
-  containerLogosButton.appendChild(imgShareImage);
-  containerLogosButton.appendChild(imgShareStickers);
-  containerLogosButton.appendChild(imgTrash);
-  containerLogosButton.appendChild(buttonPublication);
-
-  // agregando contenedores pequeños a medianos
-  sectionPublication.appendChild(figureSection);
-  sectionPublication.appendChild(inputTitle);
-  sectionPublication.appendChild(inputText);
-  sectionPublication.appendChild(containerLogosButton);
-  // agregando publicación de usuario a main
-  profileTemplate.appendChild(sectionPublication);
-
-  // divFeed.appendChild(feedTemplate);
   profileContainer.appendChild(headerTemplate());
-  profileContainer.appendChild(profileTemplate);
-  profileContainer.appendChild(textProfile);
-
-  buttonPublication.addEventListener('click', () => {
-    dataPublication(inputTitle.value, inputText.value);
-    reviewResultPublication();
-  });
+  /*profileContainer.appendChild( );*/ // aqui iria el contenedor que tenga la foto del usuario en grande y su portada
+  profileContainer.appendChild(mainTemplate);
 
   escuchandoFondo();
   return profileContainer;
