+++ src/style.css
@@ -518,11 +518,17 @@
     margin:5% 12%;
     border-radius: 15px;
     padding:1%;
+    background-color: rgb(207 159 159 / 36%);
     
 }
 .container-publication-section {
     flex-direction: column;
-   font-size: 10px;
+    font-size: 10px;
+    margin:5% 12%;
+    border-radius: 15px;
+    padding:1%;
+    width:90%;
+    background-color: white;
     
 }
 .container-publication-section input {
@@ -651,7 +657,6 @@
         align-items: center;
         justify-content: space-around;
         margin:7% 25%;
-        
         width:50%;
     }
     .container-publication-section {
